BMP180 Sensor Library
Theese are the necessary files for the BMP180 operation through I2C
interface in most of the Atmel MCUs. If it doesn't work for your project
or it doesn't compile, check your MCU datasheet for the compatibility of
the register names i used in TWI library. The TWI library is needed for
a normal operation, otherwise the code will not sompile.